cpp-port: batch 56 - port 5 hw/ files to C++
Port to .cpp:
- hw/cpu/arm11mpcore.cpp
- hw/remote/memory.cpp
- hw/uefi/hardware-info.cpp
- hw/usb/chipidea.cpp
- hw/xtensa/xtensa_memory.cpp

Add extern "C" guards to 10 headers:
- include/hw/cpu/arm11mpcore.h
- include/hw/uefi/hardware-info.h
- include/hw/usb/chipidea.h
- include/hw/arm/aspeed_coprocessor.h, max78000_soc.h
- hw/usb/hcd-ehci.h
- hw/xtensa/xtensa_memory.h
- hw/9pfs/coth.h

C++ fixes: static_cast for void*, sign comparisons, enum
casts, uint32_t loop variable for sign comparison.

Reverted: coth.c (9p.h 'private' keyword), max78000fthr.c and
aspeed_coprocessor_common.c (per-target poisoned macros).

+++ hw/usb/chipidea.cpp
@@ -95,7 +95,7 @@ static void chipidea_init(Object *obj)
     ChipideaState *ci = CHIPIDEA(obj);
     int i;
 
-    for (i = 0; i < ARRAY_SIZE(ci->iomem); i++) {
+    for (i = 0; static_cast<size_t>(i) < ARRAY_SIZE(ci->iomem); i++) {
         const struct {
             const char *name;
             hwaddr offset;
